+++ index.js
@@ -1,6 +1,7 @@
 exports.Digraph = require("./lib/Digraph");
 exports.alg = {
   isAcyclic: require("./lib/alg/isAcyclic"),
+  dijkstra: require("./lib/alg/dijkstra"),
   findCycles: require("./lib/alg/findCycles"),
   tarjan: require("./lib/alg/tarjan"),
   topsort: require("./lib/alg/topsort")